[rdk-b][mt7986][wifi-hal][Add tri-band support]
[Description]
Add hal tri-band support, and the function transform radio index to band by checking frequency.
After this commit the hal function should consider tri-band and transform index to band if needed.
And some previous hal may need to be refactored with the change. This will be done after short-term target finished.
WiFi-agent also use hal to determine which index present in which band. WiFi-agent will still work with this change.

[Release-log]
N/A

Change-Id: Iab58c035e1a20fe7c0540dd7ddaffe1e68f7dde5