[ALPS08355853][rdkb][common][hal][fix ACS with BW=80M issue]
[Description]
Fix ACS issue.
If the bandwidth is set to 20M and then 80M after the ACS is enabled, the wifi interface init fails.
The reason is that the RDK-B WiFi HAL doesn't handle the scenario of ACS (channel=0), so the parameter ht_capab=[HT40+] is missing in hostapd.conf.

[Release-log]

Feature: Wi-Fi Driver CONNAC
Change-Id: I197ec4117c9f1b2353f443ccead4d4d5d03b5199